replacereplace language handlers

+++ core/handlers/theme_handlers.py
@@ -1,8 +1,7 @@
 import os
 
 from aiogram import Bot
-from aiogram.types import Message, CallbackQuery, FSInputFile, PollAnswer
-from aiogram.enums import ParseMode
+from aiogram.types import Message, CallbackQuery, FSInputFile
 
 from config import messages
 from core.handlers.basic import is_user_subscribed
@@ -11,98 +10,13 @@ from core.image.image_analiz import image_color_picker
 from core.keyboards import inline_keybords
 from core.keyboards.reply_keybords import user_keyboard
 from core.theme_creator import create_theme
-from core.database import add_theme_to_catalog, add_language_to_catalog
+from core.database import add_theme_to_catalog
 from core.utils import is_admin, dell_data, is_private_chat
-from core.handlers.mailing_handlers import save_media_group_post_media
 
 
 USER_DATA = {}
 ADMIN_ADD_THEME = {}
-ADMIN_ADD_LANGUAGE = {}
 
-
-async def start_add_language(message: Message, bot: Bot):
-    admin = message.from_user.id
-    if is_admin(admin):
-        await message.delete()
-        ADMIN_ADD_LANGUAGE[admin] = {'language': {}}
-        ADMIN_ADD_LANGUAGE[admin]['init'] = True
-        await message.answer_poll(
-            question=messages.MESSAGE_CHOICE_DEVICE_FOR_LANGUAGE,
-            options=messages.DEVICE_FOR_LANGUAGE,
-            is_anonymous=False,
-            allows_multiple_answers=True
-        )
-
-
-async def add_language_device(poll: PollAnswer, bot: Bot):
-    admin = poll.user.id
-    if is_admin(admin) and ADMIN_ADD_LANGUAGE.get(admin):
-        if ADMIN_ADD_LANGUAGE[admin]['init']:
-            devices = [messages.DEVICE_FOR_LANGUAGE[i] for i in poll.option_ids]
-            ADMIN_ADD_LANGUAGE[admin]['language']['devices'] = {
-                'android': 'True' if 'android' in devices else 'False',
-                'ios': 'True' if 'ios' in devices else 'False',
-                'computer': 'True' if 'computer' in devices else 'False',
-            }
-            await bot.send_message(chat_id=poll.user.id,
-                                   text=messages.MESSAGE_CHOICE_CATEGORY_FOR_LANGUAGE,
-                                   reply_markup=inline_keybords.language_categories_ikb())
-
-
-async def add_language_preview(callback_query: CallbackQuery):
-    admin = callback_query.from_user.id
-    if is_admin(admin) and ADMIN_ADD_LANGUAGE.get(admin):
-        if ADMIN_ADD_LANGUAGE[admin]['init']:
-            ADMIN_ADD_LANGUAGE[admin]['language']['category'] = callback_query.data.split('_')[-1]
-            await callback_query.message.answer(text=messages.MESAGE_SEND_ME_PREVIEW_AND_TEXT, parse_mode=ParseMode.HTML)
-
-
-async def add_previev_and_desc_for_language(message: Message, bot: Bot):
-    admin = message.from_user.id
-    if is_admin(admin) and ADMIN_ADD_LANGUAGE.get(admin):
-        if ADMIN_ADD_LANGUAGE[admin]['init']:
-            media_group_id = message.media_group_id
-            preview = message.photo[-1].file_id
-            description = message.caption[5:]
-            ADMIN_ADD_LANGUAGE[admin]['language']['preview'] = [preview]
-            ADMIN_ADD_LANGUAGE[admin]['language']['description'] = description
-            ADMIN_ADD_LANGUAGE[admin]['media_group_id'] = media_group_id
-
-
-async def save_language_to_db(message, admin):
-    android = ADMIN_ADD_LANGUAGE[admin]['language']['devices']['android']
-    ios = ADMIN_ADD_LANGUAGE[admin]['language']['devices']['ios']
-    computer = ADMIN_ADD_LANGUAGE[admin]['language']['devices']['computer']
-    category = ADMIN_ADD_LANGUAGE[admin]['language']['category']
-    preview = ', '.join(ADMIN_ADD_LANGUAGE[admin]['language']['preview'])
-    description = ADMIN_ADD_LANGUAGE[admin]['language']['description']
-    await add_language_to_catalog(
-        android=android,
-        ios=ios,
-        computer=computer,
-        category=category,
-        preview=preview,
-        description=description
-    )
-    await message.answer(text=messages.MESSAGE_LANGUAGE_IS_SAVE)
-
-
-async def add_preview_for_language(message: Message, bot: Bot):
-    admin = message.from_user.id
-    media_group_id = message.media_group_id
-    try:
-        if ADMIN_ADD_LANGUAGE[admin]['media_group_id'] == media_group_id:
-            if is_admin(admin) and ADMIN_ADD_LANGUAGE.get(admin):
-                if ADMIN_ADD_LANGUAGE[admin]['init']:
-                    ADMIN_ADD_LANGUAGE[admin]['language']['preview'].append(message.photo[-1].file_id)
-                if len(ADMIN_ADD_LANGUAGE[admin]['language']['preview']) == 3:
-                    await save_language_to_db(message, admin)
-                        
-        else:
-            await save_media_group_post_media(message)
-    except:
-        await save_media_group_post_media(message)
 
         
 async def start_add_theme(message: Message, bot: Bot):
@@ -134,9 +48,16 @@ async def handler_abort(callback_query: CallbackQuery, bot: Bot):
     """ бробник натискання на кнопку відміни """
 
     chat_id = callback_query.message.chat.id
-    await callback_query.message.delete()
-    await dell_data(user_data=USER_DATA, chat_id=chat_id)
-    USER_DATA[chat_id] = {}
+    if USER_DATA[chat_id]['user_id'] == callback_query.from_user.id:
+        await callback_query.message.delete()
+        await dell_data(user_data=USER_DATA, chat_id=chat_id)
+        USER_DATA[chat_id] = {}
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
 async def handle_photo(message: Message, bot: Bot):
@@ -246,7 +167,7 @@ async def command_user_kb(message: Message, bot: Bot):
         await message.answer(text=messages.MESSAGE_ON_BACK_TO_USER_KB, reply_markup=user_keyboard(user_id))
 
 
-async def handler_device(callback_query: CallbackQuery):
+async def handler_device(callback_query: CallbackQuery, bot: Bot):
     """ Обробник вибору девайса """
     
     chat_id = callback_query.message.chat.id
@@ -258,9 +179,15 @@ async def handler_device(callback_query: CallbackQuery):
             caption=messages.CHOOSE_BACKGROUND_COLOR_TEXT,
             reply_markup=inline_keybords.choose_background_color_keyboard(USER_DATA[chat_id]['colors'])
         )
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_back_to_device_choose(callback_query: CallbackQuery):
+async def handler_back_to_device_choose(callback_query: CallbackQuery, bot: Bot):
     """ Обробник повернення до вибору девайса """
     chat_id = callback_query.message.chat.id
     if USER_DATA[chat_id]['user_id'] == callback_query.from_user.id:
@@ -268,9 +195,15 @@ async def handler_back_to_device_choose(callback_query: CallbackQuery):
             caption=messages.CHOOSE_DEVICE_TEXT,
             reply_markup=inline_keybords.choose_device_keyboard()
         )
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_background_color(callback_query: CallbackQuery):
+async def handler_background_color(callback_query: CallbackQuery, bot: Bot):
     """ Обробник вибору кольору фону """
 
     chat_id = callback_query.message.chat.id
@@ -282,9 +215,15 @@ async def handler_background_color(callback_query: CallbackQuery):
             caption=messages.CHOOSE_PRIMARY_COLOR_TEXT,
             reply_markup=inline_keybords.choose_primary_text_color_keyboard(USER_DATA[chat_id]['colors'])
         )
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_back_to_background_color_choose(callback_query: CallbackQuery):
+async def handler_back_to_background_color_choose(callback_query: CallbackQuery, bot: Bot):
     """ Обробник повернення до вибору кольору фону """
 
     chat_id = callback_query.message.chat.id
@@ -293,9 +232,15 @@ async def handler_back_to_background_color_choose(callback_query: CallbackQuery)
             caption=messages.CHOOSE_BACKGROUND_COLOR_TEXT,
             reply_markup=inline_keybords.choose_background_color_keyboard(USER_DATA[chat_id]['colors'])
         )
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_primary_text_color(callback_query: CallbackQuery):
+async def handler_primary_text_color(callback_query: CallbackQuery, bot: Bot):
     """ Обробник вибору кольору основного тексту """
 
     chat_id = callback_query.message.chat.id
@@ -307,9 +252,15 @@ async def handler_primary_text_color(callback_query: CallbackQuery):
             caption=messages.CHOOSE_SECONDARY_COLOR_TEXT,
             reply_markup=inline_keybords.choose_secondary_text_color_keyboard(USER_DATA[chat_id]['colors'])
         )
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_back_to_primary_text_color_choose(callback_query: CallbackQuery):
+async def handler_back_to_primary_text_color_choose(callback_query: CallbackQuery, bot: Bot):
     """ Обробник повернення до вибору кольору основного тексту """
 
     chat_id = callback_query.message.chat.id
@@ -318,9 +269,15 @@ async def handler_back_to_primary_text_color_choose(callback_query: CallbackQuer
             caption=messages.CHOOSE_PRIMARY_COLOR_TEXT,
             reply_markup=inline_keybords.choose_primary_text_color_keyboard(USER_DATA[chat_id]['colors'])
         )
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_secondary_text_color(callback_query: CallbackQuery):
+async def handler_secondary_text_color(callback_query: CallbackQuery, bot: Bot):
     """ Обробник вибору кольору не основного тексту """
 
     chat_id = callback_query.message.chat.id
@@ -340,9 +297,15 @@ async def handler_secondary_text_color(callback_query: CallbackQuery):
                                                                     caption=messages.MESSAGE_THEME_DONE)
             await dell_data(user_data=USER_DATA, chat_id=chat_id)
             USER_DATA[chat_id] = {}
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_back_to_secondary_text_color_choose(callback_query: CallbackQuery):
+async def handler_back_to_secondary_text_color_choose(callback_query: CallbackQuery, bot: Bot):
     """ Обробник повернення до вибору кольору не основного тексту """
 
     chat_id = callback_query.message.chat.id
@@ -351,25 +314,38 @@ async def handler_back_to_secondary_text_color_choose(callback_query: CallbackQu
             caption=messages.CHOOSE_SECONDARY_COLOR_TEXT,
             reply_markup=inline_keybords.choose_secondary_text_color_keyboard(USER_DATA[chat_id]['colors'])
         )
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_alfa_background_color(callback_query: CallbackQuery):
+async def handler_alfa_background_color(callback_query: CallbackQuery, bot: Bot):
     """ Обробник вибору прозорості для повідомлення """
 
     chat_id = callback_query.message.chat.id
-    if USER_DATA[chat_id]['device'] != 'iphone' and USER_DATA[chat_id]['user_id'] == callback_query.from_user.id:
-        background_alfa = callback_query.data.split('_')[-1]
-        USER_DATA[chat_id]['background_alfa'] = background_alfa
+    if USER_DATA[chat_id]['user_id'] == callback_query.from_user.id:
+        if USER_DATA[chat_id]['device'] != 'iphone':
+            background_alfa = callback_query.data.split('_')[-1]
+            USER_DATA[chat_id]['background_alfa'] = background_alfa
 
-        theme = await create_theme(USER_DATA[chat_id], chat_id)
-        await callback_query.message.delete()
-        await USER_DATA[chat_id]['sended_photo'].reply_document(document=FSInputFile(path=theme),
-                                                                caption=messages.MESSAGE_THEME_DONE)
-        await dell_data(user_data=USER_DATA, chat_id=chat_id)
-        USER_DATA[chat_id] = {}
+            theme = await create_theme(USER_DATA[chat_id], chat_id)
+            await callback_query.message.delete()
+            await USER_DATA[chat_id]['sended_photo'].reply_document(document=FSInputFile(path=theme),
+                                                                    caption=messages.MESSAGE_THEME_DONE)
+            await dell_data(user_data=USER_DATA, chat_id=chat_id)
+            USER_DATA[chat_id] = {}
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
 
 
-async def handler_auto_theme(callback_query: CallbackQuery):
+async def handler_auto_theme(callback_query: CallbackQuery, bot: Bot):
     """ Обробник автоматичного вибору кольору для теми """
 
     chat_id = callback_query.message.chat.id
@@ -414,3 +390,9 @@ async def handler_auto_theme(callback_query: CallbackQuery):
                                                                     caption=messages.MESSAGE_THEME_DONE)
             await dell_data(user_data=USER_DATA, chat_id=chat_id)
             USER_DATA[chat_id] = {}
+    else:
+        await bot.answer_callback_query(
+            callback_query_id=callback_query.id,
+            text=messages.MESSAGE_IS_NOT_YOUR_THEME,
+            show_alert=True
+        )
